# Connector/tt/trouble-ticketing/docs/swagger.yaml
definitions:
  handler.CreatePersonParamsArgs:
    properties:
      category:
        description: 总类
        type: integer
      group_user:
        description: 组员
        type: integer
      impact:
        description: 影响级别
        type: integer
      item:
        description: 名目
        type: integer
      target_user:
        description: 目标用户。可以为空。如果为空，在前端为"默认系统配置"
        type: string
      type:
        description: 子类
        type: integer
      work_group:
        description: 工作组
        type: integer
    required:
    - category
    - group_user
    - impact
    - item
    - type
    - work_group
    type: object
  handler.CreatePersonParamsByCopyArgs:
    properties:
      target_user:
        type: string
      template_user:
        type: string
    required:
    - target_user
    - template_user
    type: object
  model.PersonPreConfigParams:
    properties:
      category:
        description: 总类
        type: integer
      edit_user:
        type: string
      group_user:
        description: 组员
        type: integer
      id:
        type: integer
      impact:
        description: 影响级别
        type: integer
      item:
        description: 名目
        type: integer
      target_user:
        description: 目标用户。可以为空。如果为空，在前端为"默认系统配置"
        type: string
      type:
        description: 子类
        type: integer
      work_group:
        description: 工作组
        type: integer
    required:
    - category
    - group_user
    - impact
    - item
    - type
    - work_group
    type: object
  model.Ticket:
    properties:
      apply_user:
        type: string
      category:
        type: integer
      closed_time:
        type: string
      content:
        type: string
      created_at:
        type: string
      email_list:
        type: string
      group_user:
        type: integer
      id:
        type: integer
      impact:
        type: integer
      item:
        type: integer
      "no":
        type: string
      one_time_resolve_rate:
        type: integer
      resolve_cost:
        type: integer
      resolve_time:
        type: string
      resolve_timeout_sent:
        type: integer
      response_cost:
        type: integer
      response_time:
        type: string
      response_timeout_sent:
        type: integer
      root_cause:
        type: string
      solution:
        type: string
      status:
        type: string
      submit_user:
        type: string
      title:
        type: string
      type:
        type: integer
      workgroup:
        type: integer
    type: object
info:
  contact: {}
paths:
  /person/create:
    post:
      consumes:
      - application/json
      description: 创建运维人员预配置
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePersonParamsArgs'
      responses:
        "200":
          description: OK
      summary: 创建运维人员预配置
      tags:
      - 运维预配置
  /person/create/by_copy:
    post:
      consumes:
      - application/json
      description: 通过复制账号创建运维人员预配置
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePersonParamsByCopyArgs'
      responses:
        "200":
          description: OK
      summary: 通过复制账号创建运维人员预配置
      tags:
      - 运维预配置
  /person/delete/:id:
    delete:
      description: 根据TargetUser的id删除指定的账号
      responses:
        "200":
          description: OK
      summary: 根据TargetUser的id删除指定的账号
      tags:
      - 运维预配置
  /person/list:
    get:
      consumes:
      - application/json
      description: 获取所有运维人员预配置
      responses: {}
      summary: 获取所有运维人员预配置
      tags:
      - 运维预配置
  /person/update:
    post:
      consumes:
      - application/json
      description: 更新运维预配置
      parameters:
      - description: 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.PersonPreConfigParams'
      responses:
        "200":
          description: OK
      summary: 更新运维预配置
      tags:
      - 运维预配置
  /statistics/get_todo_tts:
    get:
      consumes:
      - application/json
      description: 获取待办tt列表
      parameters:
      - description: 起始时间戳. 秒数
        in: query
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: query
        name: end
        required: true
        type: integer
      - description: '是否获取所有待办. 1: 获取所有待办;  0: 获取个人待办'
        in: query
        name: all
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Ticket'
            type: array
      summary: 获取待办tt列表
      tags:
      - tt统计
  /statistics/get_tts:
    get:
      consumes:
      - application/json
      description: 获取tt列表
      parameters:
      - description: 起始时间戳. 秒数
        in: path
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: path
        name: end
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Ticket'
            type: array
      summary: 获取tt列表
      tags:
      - tt统计
  /statistics/get_user_accounts:
    get:
      consumes:
      - application/json
      description: 获取所有运维人员预配置
      parameters:
      - description: 起始时间戳. 秒数
        in: path
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: path
        name: end
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      summary: 获取所有运维人员预配置
      tags:
      - tt统计
  /statistics/work_order_summary:
    get:
      consumes:
      - application/json
      description: 获取工单按类别统计
      parameters:
      - description: 起始时间戳. 秒数
        in: query
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: query
        name: end
        required: true
        type: integer
      responses: {}
      summary: 获取工单按类别统计
      tags:
      - tt统计
  /statistics/work_order_summary/by_apply_user:
    get:
      consumes:
      - application/json
      description: 获取工单按申请人统计(只返回前10)
      parameters:
      - description: 起始时间戳. 秒数
        in: query
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: query
        name: end
        required: true
        type: integer
      responses: {}
      summary: 获取工单按申请人统计(只返回前10)
      tags:
      - tt统计
  /statistics/work_order_summary/by_status:
    get:
      consumes:
      - application/json
      description: 获取工单按 待办/完成 统计(只返回前10)
      parameters:
      - description: 起始时间戳. 秒数
        in: query
        name: start
        required: true
        type: integer
      - description: 结束时间戳. 秒数
        in: query
        name: end
        required: true
        type: integer
      - description: '是否结束。1: 完成的工单; 0: 未完成工单'
        in: query
        name: status
        required: true
        type: integer
      responses: {}
      summary: 获取工单按 待办/完成 统计(只返回前10)
      tags:
      - tt统计
swagger: "2.0"